fix(mock-doc): implement  getElementById() in document fragment (#2032)
fixes #2030
